fix: changed package name for tests
