+++ samples/snippets/src/test/java/video/DetectLogoTest.java
@@ -14,20 +14,20 @@
  * limitations under the License.
  */
 
-package com.example.video;
+package video;
 
 import static com.google.common.truth.Truth.assertThat;
 
-import video.DetectPerson;
 import java.io.ByteArrayOutputStream;
+import java.io.IOException;
 import java.io.PrintStream;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeoutException;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-@SuppressWarnings("checkstyle:AbbreviationAsWordInName")
-public class DetectPersonIT {
-
+public class DetectLogoTest {
   private ByteArrayOutputStream bout;
   private PrintStream out;
   private PrintStream originalPrintStream;
@@ -48,9 +48,14 @@ public class DetectPersonIT {
   }
 
   @Test
-  public void testDetectPerson() throws Exception {
-    DetectPerson.detectPerson("resources/googlework_short.mp4");
+  public void testLogoDetect()
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
+    LogoDetection.detectLogo("resources/googlework_short.mp4");
     String got = bout.toString();
-    assertThat(got).contains("Landmark");
+
+    assertThat(got).contains("Description");
+    assertThat(got).contains("Confidence");
+    assertThat(got).contains("Start Time Offset");
+    assertThat(got).contains("End Time Offset");
   }
 }
